Экспорт проекта в альбом пользователя с указанием идентификатора пользователя

+++ src/main/java/com/maxifly/vapi/model/AlbumAddrParseResult.java
@@ -4,8 +4,8 @@ package com.maxifly.vapi.model;
  * Created by Maximus on 08.12.2016.
  */
 public class AlbumAddrParseResult {
-    public int ownerId;
-    public int albumId;
+    public long ownerId;
+    public long albumId;
 
     public AlbumAddrParseResult(int ownerId, int albumId) {
         this.ownerId = ownerId;
